requirements.txt
albumentations==1.4.10
ConfigArgParse==1.7
numpy==1.24.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.4.0
pycm==4.0
pytorch_lightning==2.3.1
timm==0.4.12
torch==2.3.0
torch_geometric==2.6.0
torchmetrics==1.4.0.post0
torchsummary==1.5.1
torchvision==0.18.0
evalutils==0.4.2
opencv-python==4.10.0.84
scikit-learn==1.5.1
scipy==1.14.0